+++ src/onegov.ballot/HISTORY.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-- Allows to group ballot results by districts.
+- Optimizes the ballot model.
   [msom]
 
 3.4.1 (2018-03-29)